Uploaded Day3 Tasks

+++ Exercises/Day2/dynamicRouteServer.js
@@ -15,15 +15,17 @@ const server = http.createServer((req, res) => {
         res.end('Welcome to the Node.js Tutorial');
     }
 
+    //list PATH:
     else if (url.pathname === '/list') {
-        const location = path.join(__dirname, "files");
+        // const location = path.join(__dirname, "files");
 
+        //CHECK IF THE FOLDER "FILE" EXISTS OR NOT
         if (!fs.existsSync(location)) {
             res.writeHead(500, { 'Content-Type': 'text/plain' });
             return res.end('no folder called files exists.');
         }
         else {
-            fs.readdir(location, (err, files) => {
+            fs.readdir(location, { withFileTypes: true }, (err, files) => {
                 if (err) {
                     res.statusCode = 500;
                     res.setHeader('Content-Type', 'text/plain');
@@ -32,18 +34,20 @@ const server = http.createServer((req, res) => {
                 else {
                     res.statusCode = 200;
                     res.setHeader('Content-Type', 'application/json');
-    
-                    if (files.length === 0) {
+                    //FILTER FILES ONLY
+                    const fileList = files.filter(file => file.isFile()).map(file => file.name);
+
+                    //ERROR IF NO ANY FILES FOUND
+                    if (fileList.length === 0) {
                         return res.end(JSON.stringify({ message: "No files found" }));
                     }
-    
-                    return res.end(JSON.stringify(files));
+                    return res.end(JSON.stringify(fileList));
                 }
-            });
+            })
         }
 
     }
-
+    //file PATH:
     else if (url.pathname === '/file') {
         const fileName = url.searchParams.get('name');
         console.log(fileName);
@@ -76,67 +80,91 @@ const server = http.createServer((req, res) => {
         });
 
     }
-
+    //create PATH
     else if (url.pathname === '/create') {
-        const fileName = url.searchParams.get('name');
+        let fileName = url.searchParams.get('name');
         const fileContent = url.searchParams.get('content');
 
+        //IF FOLDER "FILE" DOESN'T ALREADY EXISTS, WE CREATE THE FOLDER FIRST AND THEN CRTEATE THE FILE
         if (!fs.existsSync(location)) {
             fs.mkdirSync(location);
             console.log("Folder 'files' created.");
         } else {
             console.log("Folder 'files' already exists.");
         }
+
+        //QUERY PARAMETER VALIDATION
         if (!fileName || !fileContent) {
             res.writeHead(400, { 'Content-Type': 'text/plain' });
             return res.end('ERROR: filename & content query parameters both are required');
         }
 
-        if (!fileName.endsWith(".txt")) {
-            res.writeHead(400, { 'Content-Type': 'text/plain' });
-            return res.end('ERROR: write appropriate extension i.e. .txt');
-        }
-
+        //CHECKS FOR INVALID FILE NAME 
         const invalidChars = /[<>:"/\\|?*]/;
         if (invalidChars.test(fileName)) {
             res.writeHead(400, { 'Content-Type': 'application/json' });
             return res.end(JSON.stringify({ error: 'Invalid characters in filename' }));
         }
 
-        if (fs.existsSync(fileName)) {
-            res.writeHead(400, { 'Content-Type': 'text/plain' });
-            return res.end(`ERROR: file with file name ${fileName} already exists. `);
-        } 
-        else {
-            const filePath = path.join(location, fileName);
-
-            fs.writeFile(filePath, fileContent, err => {
-                if (err) {
-                    res.writeHead(500, { 'Content-Type': 'text/plain' });
-                    return res.end('Error creating the file');
-                } else {
-                    res.writeHead(200, { 'Content-Type': 'text/plain' });
-                    res.end(`File ${fileName} created successfully with content: "${fileContent}"`);
+        //CHECKS IF USER HAS GIVEN ANY FILE EXTENSION
+        const regex = /\.[a-zA-Z0-9]+$/;
+        if (regex.test(fileName)) {
+            //IF GIVEN, THEN IT ACCEPTS ONLY .txt EXTENSION OTHERWISE THROWS ERROR
+            if (!fileName.endsWith(".txt")) {
+                res.writeHead(400, { 'Content-Type': 'text/plain' });
+                return res.end('ERROR: write appropriate extension i.e. .txt');
+            }
+            //IF NOT, THEN WE ADD .txt EXTENSION
+            else {
+                let fileName = fileName + ".txt";
+                //CHECKS IF FILE ALREADY EXISTS
+                if (fs.existsSync(fileName)) {
+                    res.writeHead(400, { 'Content-Type': 'text/plain' });
+                    return res.end(`ERROR: file with file name ${fileName} already exists. `);
                 }
-            });
+                else {
+                    const filePath = path.join(location, fileName);
+
+                    fs.writeFile(filePath, fileContent, err => {
+                        if (err) {
+                            res.writeHead(500, { 'Content-Type': 'text/plain' });
+                            return res.end('Error creating the file');
+                        } else {
+                            res.writeHead(200, { 'Content-Type': 'text/plain' });
+                            res.end(`File ${fileName} created successfully with content: "${fileContent}"`);
+                        }
+                    });
+                }
+            }
         }
     }
 
+    ///append PATH
     else if (url.pathname === '/append') {
         const fileName = url.searchParams.get('name');
         const fileContent = url.searchParams.get('content');
 
+        //QUERY PARAMETER VALIDATION
         if (!fileName || !fileContent) {
             res.writeHead(400, { 'Content-Type': 'text/plain' });
             return res.end('ERROR: filename & content query parameters both are required');
         }
+
+        //ACCEPTS ONLY .txt EXTENSION OTHERWISE THROWS ERROR
         if (!fileName.endsWith(".txt")) {
             res.writeHead(400, { 'Content-Type': 'text/plain' });
             return res.end('ERROR: write appropriate extension i.e. .txt');
         }
-        const filePath = path.join(location, fileName);
 
-        fs.appendFile(filePath, fileContent, err => {
+        const filePath = path.join(location, fileName);
+        
+        //CHECKS IF FILE DOESN'T EXISTS
+        if (!fs.existsSync(filePath)) {
+            res.writeHead(400, { 'Content-Type': 'text/plain' });
+            return res.end(`ERROR: file with file name ${fileName} doesn't exists. `);
+        }
+
+        fs.appendFile(filePath, `\n${fileContent}`, err => {
             if (err) {
                 res.writeHead(500, { 'Content-Type': 'text/plain' });
                 return res.end('Error appending the content in the file');
@@ -147,18 +175,27 @@ const server = http.createServer((req, res) => {
         });
     }
 
+    //delete PATH
     else if (url.pathname === '/delete') {
         const fileName = url.searchParams.get('name');
 
+        //QUERY PARAMETER VALIDATION
         if (!fileName) {
             res.writeHead(400, { 'Content-Type': 'text/plain' });
             return res.end('ERROR: filename query parameters is required');
         }
+        //ACCEPTS ONLY .txt EXTENSION OTHERWISE THROWS ERROR
         if (!fileName.endsWith(".txt")) {
             res.writeHead(400, { 'Content-Type': 'text/plain' });
             return res.end('ERROR: write appropriate extension i.e. .txt');
         }
         const filePath = path.join(location, fileName);
+
+         //CHECKS IF FILE DOESN'T EXISTS
+         if (!fs.existsSync(filePath)) {
+            res.writeHead(400, { 'Content-Type': 'text/plain' });
+            return res.end(`ERROR: file with file name ${fileName} doesn't exists. `);
+        }
 
         fs.unlink(filePath, err => {
             if (err) {
